+++ packages/core/src/index.d.ts
@@ -48,6 +48,7 @@ export interface PluginFunctions {
     openFile(file: File): void;
     rotate(rotation: number): void;
     setViewerState(viewerState: ViewerState): void;
+    switchScrollMode(scrollMode: ScrollMode): void;
     zoom(scale: number | SpecialZoomLevel): void;
 }
 export interface PluginOnDocumentLoad {